Rename LWJGL  context header
