+++ src/main/resources/lib/libJTouchBar.dylib.dSYM/libJTouchBar.dylib
